requirements.txt
Scrapy
Twisted
service-identity
PyYAML
GitPython
html2text
